requirements.txt
boto3==1.9.233
botocore==1.12.233
Django==2.2.5
django-crispy-forms==1.7.2
django-storages==1.7.2
docutils==0.15.2
gunicorn==19.9.0
jmespath==0.9.4
mypy==0.720
mypy-extensions==0.4.1
Pillow==6.1.0
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.5